+++ EssentialFeedTests/Feed API/LoadFeedFromRemoteUseCaseTests.swift
@@ -36,7 +36,7 @@ class LoadFeedFromRemoteUseCaseTests: XCTestCase {
 
         expect(sut, toCompleteWith: failure(.connectivity), when: {
             let clientError = NSError(domain: "Test", code: 0)
-            client.complete(withError: clientError)
+            client.complete(with: clientError)
         })
     }
 
@@ -171,34 +171,5 @@ class LoadFeedFromRemoteUseCaseTests: XCTestCase {
         }
         action()
         wait(for: [exp], timeout: 1.0)
-    }
-
-    private class HTTPClientSpy: HTTPClient {
-        private struct Task: HTTPClientTask {
-            func cancel() {}
-        }
-
-        var messages = [(url: URL, completion: (HTTPClient.Result) -> Void)]()
-
-        var requestedURLs: [URL] { messages.map { $0.url } }
-
-        func get(from url: URL, completion: @escaping (HTTPClient.Result) -> Void) -> HTTPClientTask {
-            messages.append((url, completion))
-            return Task()
-        }
-
-        func complete(withError error: Error, at index: Int = 0) {
-            messages[index].completion(.failure(error))
-        }
-
-        func complete(withStatusCode code: Int, data: Data, at index: Int = 0) {
-            let response = HTTPURLResponse(
-                url: requestedURLs[index],
-                statusCode: code,
-                httpVersion: nil,
-                headerFields: nil
-            )!
-            messages[index].completion(.success((data, response)))
-        }
     }
 }
